requirements.txt
python-dotenv
mysql-connector-python
openai
fastapi
uvicorn
tiktoken
langchain
openai
langchain-openai
langchain-community
langchain-huggingface
langchainhub
pinecone-client
langchain-pinecone
sentence-transformers
pypdf
youtube-transcript-api
bs4
pytube
streamlit
streamlit_chat